Handle `SqlValue.Date` correctly when populating rows in batch
PR #126 added support for `DateOnly` by changing the value contained in
the `Date` case of the `SqlValue` type into a `Choice<DateTime,
DateOnly>`. The code in `populateRow` was updated accordingly, but the
one in `populatedBatchRow` (used for transactions) was not. This causes
the following exception to be raised:

    System.InvalidCastException: Can't write CLR type
        Microsoft.FSharp.Core.FSharpChoice`2+Choice2Of2[
        System.DateTime,System.DateOnly] with handler type DateHandler

This commit updates the `populateBatchRow` function to handle the
`Choice<DateTime, DateOnly>`.